# openapi.json
{
  "components": {
    "schemas": {
...
      "FindingOut": {
        "properties": {
          "conclusion": {
            "title": "Conclusion",
            "type": "string"
          },
          "evidenceKind": {
            "default": "passage",
            "enum": [
              "passage",
              "coverage"
            ],
            "title": "Evidencekind",
            "type": "string"
          },
          "guidanceReference": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Guidancereference"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "modelConclusion": {
            "default": "",
            "title": "Modelconclusion",
            "type": "string"
          },
          "observed": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Observed"
          },
          "required": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Required"
          },
          "review": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ReviewOut"
              },
              {
                "type": "null"
              }
            ]
          },
          "spans": {
            "items": {
              "$ref": "#/components/schemas/SpanOut"
            },
            "title": "Spans",
            "type": "array"
          },
          "status": {
            "enum": [
              "pass",
              "needs_review",
              "missing",
              "unresolved"
            ],
            "title": "Status",
            "type": "string"
          },
          "statusReason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Statusreason"
          },
          "statusSource": {
            "anyOf": [
              {
                "enum": [
                  "computed_days",
                  "confirmed_days",
                  "model_hint",
                  "no_evidence",
                  "reference_check",
                  "position_check",
                  "ambiguous_fact"
                ],
                "type": "string"
              },
              {
                "const": "",
                "type": "string"
              }
            ],
            "default": "",
            "title": "Statussource"
          },
          "suggestedPosition": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Suggestedposition"
          },
          "topic": {
            "title": "Topic",
            "type": "string"
          }
        },
        "required": [
          "id",
          "topic",
          "status",
          "conclusion",
          "spans",
          "guidanceReference",
          "observed",
          "required",
          "suggestedPosition"
        ],
        "title": "FindingOut",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "RunExplanation": {
        "properties": {
          "findings": {
            "items": {
              "$ref": "#/components/schemas/ExplanationFinding"
            },
            "title": "Findings",
            "type": "array"
          },
          "proposals": {
            "items": {
              "$ref": "#/components/schemas/ModelProposal"
            },
            "title": "Proposals",
            "type": "array"
          },
          "proposalsProblem": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Proposalsproblem"
          },
          "question": {
            "title": "Question",
            "type": "string"
          },
          "reading": {
            "$ref": "#/components/schemas/ExplanationReading"
          },
          "reason": {
            "anyOf": [
              {
                "enum": [
                  "insufficient_evidence",
                  "citations_unverified",
                  "invalid_output",
                  "provider_error",
                  "internal_error"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          },
          "reconstruction": {
            "$ref": "#/components/schemas/ExplanationReconstruction"
          },
          "reproducibility": {
            "$ref": "#/components/schemas/ExplanationReproducibility"
          },
          "retrieval": {
            "items": {
              "$ref": "#/components/schemas/ExplanationCandidate"
            },
            "title": "Retrieval",
            "type": "array"
          },
          "retrievalMode": {
            "anyOf": [
              {
                "enum": [
                  "lexical_match",
                  "hybrid_match",
                  "opening_fallback"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Retrievalmode"
          },
          "runId": {
            "title": "Runid",
            "type": "string"
          },
          "stage": {
            "enum": [
              "reading",
              "finding_evidence",
              "checking",
              "verifying",
              "complete",
              "unresolved",
              "failed"
            ],
            "title": "Stage",
            "type": "string"
          }
        },
        "required": [
          "runId",
          "question",
          "stage",
          "reason",
          "reading",
          "retrieval",
          "reconstruction",
          "proposals",
          "proposalsProblem",
          "findings",
          "reproducibility"
        ],
        "title": "RunExplanation",
        "type": "object"
      },
      "RunIn": {
        "properties": {
          "documentId": {
            "title": "Documentid",
            "type": "string"
          },
          "guidanceId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Guidanceid"
          },
          "question": {
            "maxLength": 2000,
            "minLength": 1,
            "title": "Question",
            "type": "string"
          }
        },
        "required": [
          "documentId",
          "question"
        ],
        "title": "RunIn",
        "type": "object"
      },
      "RunOut": {
        "properties": {
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "findings": {
            "items": {
              "$ref": "#/components/schemas/FindingOut"
            },
            "title": "Findings",
            "type": "array"
          },
          "guidanceText": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Guidancetext"
          },
          "hasGuidance": {
            "default": false,
            "title": "Hasguidance",
            "type": "boolean"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "latencyMs": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Latencyms"
          },
          "model": {
            "title": "Model",
            "type": "string"
          },
          "note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          },
          "promptHash": {
            "title": "Prompthash",
            "type": "string"
          },
          "promptVersion": {
            "title": "Promptversion",
            "type": "string"
          },
          "question": {
            "title": "Question",
            "type": "string"
          },
          "reason": {
            "anyOf": [
              {
                "enum": [
                  "insufficient_evidence",
                  "citations_unverified",
                  "invalid_output",
                  "provider_error",
                  "internal_error"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          },
          "retrievalMode": {
            "anyOf": [
              {
                "enum": [
                  "lexical_match",
                  "hybrid_match",
                  "opening_fallback"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Retrievalmode"
          },
          "reused": {
            "default": false,
            "title": "Reused",
            "type": "boolean"
          },
          "reviewHead": {
            "default": "",
            "title": "Reviewhead",
            "type": "string"
          },
          "sectionsRead": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sectionsread"
          },
          "sectionsRequested": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sectionsrequested"
          },
          "shared": {
            "default": false,
            "title": "Shared",
            "type": "boolean"
          },
          "stage": {
            "enum": [
              "reading",
              "finding_evidence",
              "checking",
              "verifying",
              "complete",
              "unresolved",
              "failed"
            ],
            "title": "Stage",
            "type": "string"
          },
          "withheld": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/FindingOut"
            },
            "title": "Withheld",
            "type": "array"
          }
        },
        "required": [
          "id",
          "question",
          "stage",
          "findings",
          "model",
          "promptVersion",
          "promptHash",
          "latencyMs",
          "note",
          "error"
        ],
        "title": "RunOut",
        "type": "object"
      },
      "RunSummary": {
        "properties": {
          "createdAt": {
            "title": "Createdat",
            "type": "string"
          },
          "documentId": {
            "title": "Documentid",
            "type": "string"
          },
          "documentName": {
            "title": "Documentname",
            "type": "string"
          },
          "findings": {
            "title": "Findings",
            "type": "integer"
          },
          "hasGuidance": {
            "title": "Hasguidance",
            "type": "boolean"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "latencyMs": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Latencyms"
          },
          "model": {
            "title": "Model",
            "type": "string"
          },
          "promptHash": {
            "title": "Prompthash",
            "type": "string"
          },
          "promptVersion": {
            "title": "Promptversion",
            "type": "string"
          },
          "question": {
            "title": "Question",
            "type": "string"
          },
          "reason": {
            "anyOf": [
              {
                "enum": [
                  "insufficient_evidence",
                  "citations_unverified",
                  "invalid_output",
                  "provider_error",
                  "internal_error"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          },
          "shared": {
            "default": false,
            "title": "Shared",
            "type": "boolean"
          },
          "stage": {
            "enum": [
              "reading",
              "finding_evidence",
              "checking",
              "verifying",
              "complete",
              "unresolved",
              "failed"
            ],
            "title": "Stage",
            "type": "string"
          }
        },
        "required": [
          "id",
          "question",
          "stage",
          "model",
          "promptVersion",
          "promptHash",
          "latencyMs",
          "findings",
          "hasGuidance",
          "documentId",
          "documentName",
          "createdAt"
        ],
        "title": "RunSummary",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Contract Workbench",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/runs": {
      "get": {
        "description": "This workspace's runs and the curated ones, newest first, the latest 200.",
        "operationId": "list_runs_api_runs_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/RunSummary"
                  },
                  "title": "Response List Runs Api Runs Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "List Runs",
        "tags": [
          "runs"
        ]
      },
      "post": {
        "description": "202 with a new run, or 200 with this workspace's running or completed run that already has these inputs.",
        "operationId": "create_run_api_runs_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RunIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Run",
        "tags": [
          "runs"
        ]
      }
    },
    "/api/runs/{run_id}": {
      "get": {
        "operationId": "get_run_api_runs__run_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Run",
        "tags": [
          "runs"
        ]
      }
    },
    "/api/runs/{run_id}/detail": {
      "get": {
        "operationId": "get_run_detail_api_runs__run_id__detail_get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Run Detail",
        "tags": [
          "runs"
        ]
      }
    },
    "/api/runs/{run_id}/events": {
      "get": {
        "description": "Stage transitions only. Replays what already happened, then streams until the run ends, by this process's\nbus or, at the latest, one heartbeat after the record says it ended.",
        "operationId": "run_events_api_runs__run_id__events_get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Run Events",
        "tags": [
          "runs"
        ]
      }
    },
    "/api/runs/{run_id}/evidence-pack": {
      "get": {
        "description": "A zip with the original bytes, the canonical sections, the run record, every located span and a stdlib verify.py.",
        "operationId": "evidence_pack_api_runs__run_id__evidence_pack_get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Evidence Pack",
        "tags": [
          "runs"
        ]
      }
    },
    "/api/runs/{run_id}/explanation": {
      "get": {
        "description": "Why this answer: what was read, what retrieval chose, the exact model input, the model's proposal, each\nSourceMatch and the recorded policy evaluation, all from the record. Read-only; nothing is recomputed as fact.",
        "operationId": "get_run_explanation_api_runs__run_id__explanation_get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunExplanation"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Run Explanation",
        "tags": [
          "runs"
        ]
      }
    }
  }
}